Ethernet debugging snapshot To byte swap or not to byte swap, that is the question.
* Translation.cs:
* HostInterface.cs:
* MachineAddress.cs:
* Ether10MbitController.cs: The beatings will continue until packet
  exchange improves.

* PERQemu.csproj: Remove last of the old .phd images; add some network
  changes

* Network.txt: Verbiage! I should get paid by the word, man.
