+++ PERQemu/Emulator/IO/Network/MachineAddress.cs
@@ -17,8 +17,6 @@
 // along with PERQemu.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-using System;
-using System.Collections.Generic;
 using System.Net.NetworkInformation;
 
 using PERQemu.Config;
@@ -123,155 +121,5 @@ namespace PERQemu.IO.Network
 
         byte[] _mac;
         PhysicalAddress _physAddr;
-    }
-
-
-    /// <summary>
-    /// Store an Ethernet address translation so that virtual PERQs can be easily
-    /// mapped to their hosts.  Largely for stats/debugging?
-    /// </summary>
-    public class NATEntry
-    {
-        public NATEntry(PhysicalAddress host, PhysicalAddress perq)
-        {
-            Host = host;
-            Perq = perq;
-            FirstSeen = DateTime.Now;
-            LastReceived = DateTime.Now;
-            Sent = Received = 0;
-        }
-
-        public override string ToString()
-        {
-            return $"[Host: {Host}  Perq: {Perq}]";
-        }
-
-        public PhysicalAddress Host;        // Host MAC
-        public PhysicalAddress Perq;        // PERQ's MAC
-        public DateTime FirstSeen;          // Date/time mapping established
-        public DateTime LastReceived;       // Date/time last packet/update received
-        public ulong Received;              // Count Host->Perq mappings
-        public ulong Sent;                  // Count Perq->Host mappings
-    }
-
-    /// <summary>
-    /// A (very) simple Network Address Translation table, mapping a given host
-    /// MAC address to an emulated PERQ address.
-    /// </summary>
-    public class NATTable
-    {
-        public NATTable()
-        {
-            _entries = new Dictionary<PhysicalAddress, NATEntry>();
-            _perqToHost = new Dictionary<PhysicalAddress, PhysicalAddress>();
-        }
-
-        public void Reset()
-        {
-        }
-
-        /// <summary>
-        /// Flush the entire table.
-        /// </summary>
-        public void Flush()
-        {
-            _entries.Clear();
-            Log.Info(Category.Ethernet, "NAT table flushed");
-        }
-
-        /// <summary>
-        /// Add a new mapping.  We leave initialization up to the caller.
-        /// </summary>
-        public void Add(NATEntry ent)
-        {
-            // Make sure it isn't already there...
-            if (_entries.ContainsValue(ent))
-            {
-                Log.Write(Category.Ethernet, "Can't add duplicate NAT entry, ignored {0}", ent);
-                return;
-            }
-            _entries.Add(ent.Host, ent);
-
-            // Do the inverse index too
-            if (_perqToHost.ContainsKey(ent.Perq))
-            {
-                Log.Write(Category.Ethernet, "PERQ {0} already in index at different host?", ent.Perq);
-                return;
-            }
-            Log.Write(Category.Ethernet, "NAT entry added {0}", ent);
-        }
-
-        /// <summary>
-        /// Search the table to see if we've seen PERQ traffic from a given source
-        /// MAC address before.  If so, return the NAT entry; otherwise, null.
-        /// </summary>
-        public NATEntry LookupHost(PhysicalAddress host)
-        {
-            if (_entries.ContainsKey(host))
-            {
-                return _entries[host];
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Look up another PERQ (real or virtual) in the inverse map and return
-        /// the host entry for it.
-        /// </summary>
-        public NATEntry LookupPerq(PhysicalAddress perq)
-        {
-            if (_perqToHost.ContainsKey(perq))
-            {
-                return LookupHost(_perqToHost[perq]);
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Bump the access/packet count and receive time for the given host.
-        /// Quietly no-op if not in table?
-        /// </summary>
-        public void UpdateHost(PhysicalAddress host)
-        {
-            if (_entries.ContainsKey(host))
-            {
-                _entries[host].LastReceived = DateTime.Now;
-                _entries[host].Received++;
-            }
-        }
-
-        public void UpdatePerq(PhysicalAddress perq)
-        {
-            var ent = LookupPerq(perq);
-            if (ent != null)
-            {
-                ent.Sent++;
-            }
-        }
-
-        // Debugging
-        public void DumpTable()
-        {
-            if (_entries.Count == 0)
-            {
-                Console.WriteLine("  NAT table is empty.");
-                return;
-            }
-
-            // messy as hell, gotta format this up all sweetly
-            Console.WriteLine("\nNAT table:");
-            Console.WriteLine("\nHost\t\tPerq\t\tFirst seen\tLast seen\t\tSent / Rcvd");
-
-            foreach (var e in _entries.Values)
-            {
-                //var seconds = e.LastReceived - e.FirstSeen;
-                Console.WriteLine($"{e.Host}  {e.Perq}  {e.FirstSeen}  {e.LastReceived}  {e.Sent}  {e.Received}");
-            }
-        }
-
-        Dictionary<PhysicalAddress, NATEntry> _entries;
-        Dictionary<PhysicalAddress, PhysicalAddress> _perqToHost;
     }
 }
